# Implement "is_mate" method on Sequence and BytesSequence