fixes #35  

Sorry for the PR bombardment. (Again...). Anyway this PR was very easy to make (and hopefully review). This adds an is_mate method to the Sequence and BytesSequence types. This method works with self and other, where other is another object of the same type. 

I see in the Cython generated code that Cython handles this properly: A typecheck on other is performed and Cython knows self and other are structs with a name member. Also Cython properly uses `METH_O` so no argument tuple and dict unpacking is performed. 

The mate check is performed in pairedend.py now with `r1.is_mate(r2)` instead of `record_names_match(r1.name, r2.name)`. 

This saves:
- Two PyObject_GetAttribute calls. (`r1.name` and `r2.name`). These are taken directly from the struct.
- Argument tuple and dict unpacking, as there is only one argument.
- Some checks. The Sequence and BytesSequence types guarantee their members are unicode or bytes respectively. 

This causes a significant speedup of paired-end reading:
Before (Current main):
```
Benchmark #1: python dnaio_read_paired.py ~/test/big2.fastq
  Time (mean ± σ):      2.768 s ±  0.021 s    [User: 2.560 s, System: 0.208 s]
  Range (min … max):    2.742 s …  2.800 s    10 runs
 
Benchmark #1: python dnaio_read_paired_binary.py ~/test/big2.fastq
  Time (mean ± σ):      2.561 s ±  0.027 s    [User: 2.366 s, System: 0.195 s]
  Range (min … max):    2.511 s …  2.600 s    10 runs
```

After:
```
Benchmark #1: python dnaio_read_paired.py ~/test/big2.fastq
  Time (mean ± σ):      2.556 s ±  0.016 s    [User: 2.346 s, System: 0.209 s]
  Range (min … max):    2.529 s …  2.584 s    10 runs
 
Benchmark #1: python dnaio_read_paired_binary.py ~/test/big2.fastq
  Time (mean ± σ):      2.314 s ±  0.027 s    [User: 2.117 s, System: 0.196 s]
  Range (min … max):    2.282 s …  2.369 s    10 runs
```

Together with #39 this reduces the overhead of paired-end reading significantly.