@rhpvorderman 2022-02-14
@marcelm. This PR is very tiny, but take your time. Paired-end reading still has some overhead, but at least it is at an acceptable level now (20%ish).