@rhpvorderman 2022-02-15
@marcelm done!
The error messages are slightly hypocrite, as we allow basically any latin-1 encoded character as well, not just ASCII. We have to though, as the Sequence objects we produce all have a maxchar value of 255 so all our objects would fail this check if we would be more stringent. 
See #40